update map: update maps: minor fixes

+++ code/api/map.go
@@ -63,23 +63,23 @@ type MapResponse struct {
 // Структура карты
 type Maps struct {
 	BeatmapSetId     int       `json:"beatmapset_id"`
-	DifficultyRating float32   `json:"difficulty_rating"`
+	DifficultyRating float64   `json:"difficulty_rating"`
 	Id               int       `json:"id"`
 	Mode             string    `json:"mode"`
 	Status           string    `json:"status"`
 	TotalLength      int       `json:"total_length"`
 	UserId           int       `json:"user_id"`
 	Version          string    `json:"version"`
-	Accuracy         float32   `json:"accuracy"`
-	Ar               float32   `json:"ar"`
-	Bpm              float32   `json:"bpm"`
+	Accuracy         float64   `json:"accuracy"`
+	Ar               float64   `json:"ar"`
+	Bpm              float64   `json:"bpm"`
 	Convert          bool      `json:"convert"`
 	CountCircles     int       `json:"count_circles"`
 	CountSliders     int       `json:"count_sliders"`
 	CountSpinners    int       `json:"count_spinners"`
-	Cs               float32   `json:"cs"`
+	Cs               float64   `json:"cs"`
 	DeletedAt        string    `json:"deleted_at"`
-	Drain            float32   `json:"drain"`
+	Drain            float64   `json:"drain"`
 	HitLength        int       `json:"hit_length"`
 	IsScoreable      bool      `json:"is_scoreable"`
 	LastUpdated      string    `json:"last_updated"`
